@@ -6,15 +6,18 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.8.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
+  injectable: ^2.7.0
+  flutter_secure_storage: ^9.2.4
+  freezed_annotation: ^3.1.0
+  json_annotation: ^4.9.0
   get_it: ^9.1.1
-  mobx: ^2.2.0+1
-  flutter_mobx: ^2.1.0+1
+  mobx: ^2.5.0
+  flutter_mobx: ^2.3.0
   go_router: ^17.0.0
   dio: ^5.7.0
   shared_preferences: ^2.3.3
@@ -22,19 +25,34 @@ dependencies:
   flutter_svg: ^2.0.15
   country_picker: ^2.0.21
   cupertino_icons: ^1.0.8
+  easy_localization: ^3.0.8
+  pinput: ^6.0.1
+  collection: ^1.19.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
   flutter_lints: ^6.0.0
   build_runner: ^2.4.13
-  mobx_codegen: ^2.5.0+1
+  environment_config: ^4.0.0+1
+  injectable_generator: ^2.9.1
+  mobx_codegen: ^2.7.0
+  json_serializable: ^6.11.2
+  freezed: ^3.2.3
+  flutter_gen_runner: ^5.12.0
+
+flutter_gen:
+  output: lib/core/assets/gen
+  integrations:
+    image: true
+    flutter_svg: true
 
 flutter:
   uses-material-design: true
 
   assets:
+    - assets/lang/
+    - assets/icons/
     - assets/profile/
     - assets/profile/bussines/
     - assets/bottom/
@@ -44,3 +62,13 @@ flutter:
     - family: AppIcons
       fonts:
         - asset: assets/fonts/AppIcons.ttf
+    - family: Inter
+      fonts:
+        - asset: assets/fonts/Inter-Regular.ttf
+          weight: 400
+        - asset: assets/fonts/Inter-Medium.ttf
+          weight: 500
+        - asset: assets/fonts/Inter-SemiBold.ttf
+          weight: 600
+        - asset: assets/fonts/Inter-Bold.ttf
+          weight: 700
