@@ -1,40 +1,42 @@
+import 'package:easy_localization/easy_localization.dart';
+import 'package:exchats/core/assets/gen/assets.gen.dart';
+import 'package:exchats/core/assets/gen/fonts.gen.dart';
+import 'package:exchats/features/auth/presentation/store/login_store.dart';
+import 'package:exchats/generated/locale_keys.g.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_svg/flutter_svg.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:go_router/go_router.dart';
 import 'package:country_picker/country_picker.dart';
 import 'package:exchats/core/di/locator.dart';
 import 'package:exchats/core/constants/app_colors.dart';
-import 'package:exchats/features/auth/presentation/store/auth_store.dart';
 
-class LoginScreen extends StatefulWidget {
-  const LoginScreen({Key? key}) : super(key: key);
+final class LoginScreen extends StatefulWidget {
+  const LoginScreen({super.key});
 
   @override
-  _LoginScreenState createState() => _LoginScreenState();
+  State<LoginScreen> createState() => _LoginScreenState();
 }
 
-class _LoginScreenState extends State<LoginScreen> {
-  int _selectedTab = 0;
-  final TextEditingController _phoneController = TextEditingController();
-  final TextEditingController _emailController = TextEditingController();
-  bool _isPhoneValid = false;
-  bool _isEmailValid = false;
-  String? _phoneError;
-  String? _emailError;
-  bool _isSendingCode = false;
-  late Country _selectedCountry;
+final class _LoginScreenState extends State<LoginScreen> {
+  late LoginStore store;
+  late TextEditingController phoneController;
+  late TextEditingController emailController;
 
   @override
   void initState() {
     super.initState();
-    _selectedCountry = CountryParser.parseCountryCode('RU');
+    store = locator<LoginStore>();
+    store.setupValidations();
+    phoneController = TextEditingController();
+    emailController = TextEditingController();
   }
 
   @override
   void dispose() {
-    _phoneController.dispose();
-    _emailController.dispose();
+    phoneController.dispose();
+    emailController.dispose();
+    store.dispose();
     super.dispose();
   }
 
@@ -43,181 +45,139 @@ class _LoginScreenState extends State<LoginScreen> {
     final theme = Theme.of(context);
 
     return Scaffold(
-      backgroundColor: AppColors.background,
+      backgroundColor: AppColors.surface,
       body: SafeArea(
         child: Center(
-          child: SingleChildScrollView(
-            child: Container(
-              margin: const EdgeInsets.symmetric(horizontal: 20.0),
-              padding: const EdgeInsets.all(24.0),
-              decoration: BoxDecoration(
-                color: Colors.white,
-                borderRadius: BorderRadius.circular(8.0),
-                boxShadow: [
-                  BoxShadow(
-                    color: Colors.black.withOpacity(0.1),
-                    blurRadius: 10.0,
-                    offset: Offset(0, 4),
+          child: Observer(
+            builder: (context) {
+              return SingleChildScrollView(
+                child: Container(
+                  margin: const EdgeInsets.symmetric(horizontal: 12.0),
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: 12.0,
+                    vertical: 20.0,
                   ),
-                ],
-              ),
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Image.asset(
-                    'assets/auth/logo.png',
-                    width: 64.0,
-                    height: 64.0,
-                    errorBuilder: (context, error, stackTrace) {
-                      return Container(
-                        width: 64.0,
-                        height: 64.0,
+                  decoration: BoxDecoration(
+                    color: AppColors.subSurface,
+                    borderRadius: BorderRadius.circular(12.0),
+                    border: Border.all(color: AppColors.borderSecondary),
+                  ),
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      Assets.auth.logo.image(width: 64.0, height: 64.0),
+                      const SizedBox(height: 12.0),
+                      Container(
+                        padding: const EdgeInsets.all(12.0),
                         decoration: BoxDecoration(
-                          color: Colors.grey[300],
-                          shape: BoxShape.circle,
+                          color: AppColors.onPrimary,
+                          borderRadius: BorderRadius.circular(12.0),
                         ),
-                        child: Icon(Icons.image, size: 32.0),
-                      );
-                    },
-                  ),
-                  const SizedBox(height: 24.0),
-                  Text(
-                    'Авторизация',
-                    style: TextStyle(
-                      fontSize: 24.0,
-                      fontWeight: FontWeight.bold,
-                      color: Colors.black,
-                    ),
-                  ),
-                  const SizedBox(height: 32.0),
-                  _buildTabs(theme),
-                  const SizedBox(height: 24.0),
-                  _selectedTab == 0
-                      ? _buildPhoneInput(theme)
-                      : _buildEmailInput(theme),
-                  if ((_selectedTab == 0 && _phoneError != null) ||
-                      (_selectedTab == 1 && _emailError != null))
-                    Padding(
-                      padding: const EdgeInsets.only(top: 8.0, left: 0),
-                      child: Align(
-                        alignment: Alignment.centerLeft,
-                        child: Text(
-                          _selectedTab == 0 ? _phoneError! : _emailError!,
-                          style: TextStyle(
-                            color: Colors.red,
-                            fontSize: 12.0,
-                          ),
+                        child: Column(
+                          children: [
+                            Text(
+                              LocaleKeys.auth_login_title.tr(),
+                              style: TextStyle(
+                                fontFamily: FontFamily.inter,
+                                fontSize: 24.0,
+                                fontWeight: FontWeight.w600,
+                                color: AppColors.onSurface,
+                              ),
+                            ),
+                            const SizedBox(height: 12.0),
+                            _buildTabs(),
+                            const SizedBox(height: 14.0),
+                            store.selectedTab == 0
+                                ? _buildPhoneInput(theme)
+                                : _buildEmailInput(theme),
+                            if (store.hasError)
+                              Padding(
+                                padding: const EdgeInsets.only(
+                                  top: 8.0,
+                                  left: 0,
+                                ),
+                                child: Align(
+                                  alignment: Alignment.centerLeft,
+                                  child: Text(
+                                    store.error!,
+                                    style: TextStyle(
+                                      color: Colors.red,
+                                      fontSize: 12.0,
+                                    ),
+                                  ),
+                                ),
+                              ),
+                            const SizedBox(height: 12.0),
+                            _buildNextButton(theme),
+                          ],
                         ),
                       ),
-                    ),
-                  const SizedBox(height: 24.0),
-                  _buildNextButton(theme),
-                  const SizedBox(height: 24.0),
-                  _buildSocialLogin(theme),
-                ],
-              ),
-            ),
+                      const SizedBox(height: 12.0),
+                      _buildSocialLogin(theme),
+                    ],
+                  ),
+                ),
+              );
+            },
           ),
         ),
       ),
     );
   }
 
-  Widget _buildTabs(ThemeData theme) {
+  Widget _buildTabs() {
+    Widget tab(String title, int index) {
+      final selected = store.selectedTab == index;
+      return Expanded(
+        child: GestureDetector(
+          onTap: () => store.setSelectedTab(index),
+          child: Container(
+            height: 32,
+            alignment: Alignment.center,
+            decoration: BoxDecoration(
+              color: selected ? AppColors.onPrimary : Colors.transparent,
+              borderRadius: BorderRadius.circular(6),
+            ),
+            child: Text(
+              title,
+              style: TextStyle(
+                fontFamily: FontFamily.inter,
+                fontSize: 13,
+                fontWeight: FontWeight.w400,
+                color: AppColors.onSurface,
+              ),
+            ),
+          ),
+        ),
+      );
+    }
+
     return Container(
-      height: 44.0,
+      padding: const EdgeInsets.all(2),
       decoration: BoxDecoration(
-        color: AppColors.background,
-        borderRadius: BorderRadius.circular(8.0),
+        color: AppColors.gray2.withValues(alpha: 0.08),
+        borderRadius: BorderRadius.circular(8),
       ),
       child: Row(
         children: [
-          Expanded(
-            child: GestureDetector(
-              onTap: () {
-                setState(() {
-                  _selectedTab = 0;
-                  _phoneError = null;
-                });
-              },
-              child: Container(
-                decoration: BoxDecoration(
-                  color: _selectedTab == 0 ? Colors.white : Colors.transparent,
-                  borderRadius: BorderRadius.only(
-                    topLeft: Radius.circular(8.0),
-                    bottomLeft: Radius.circular(8.0),
-                  ),
-                  border: _selectedTab == 0
-                      ? Border.all(color: Colors.grey[300]!, width: 1.0)
-                      : null,
-                ),
-                child: Center(
-                  child: Text(
-                    'Телефон',
-                    style: TextStyle(
-                      fontSize: 16.0,
-                      fontWeight: _selectedTab == 0
-                          ? FontWeight.w600
-                          : FontWeight.normal,
-                      color: Colors.black,
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ),
-          Expanded(
-            child: GestureDetector(
-              onTap: () {
-                setState(() {
-                  _selectedTab = 1;
-                  _emailError = null;
-                });
-              },
-              child: Container(
-                decoration: BoxDecoration(
-                  color: _selectedTab == 1 ? Colors.white : Colors.transparent,
-                  borderRadius: BorderRadius.only(
-                    topRight: Radius.circular(8.0),
-                    bottomRight: Radius.circular(8.0),
-                  ),
-                  border: _selectedTab == 1
-                      ? Border.all(color: Colors.grey[300]!, width: 1.0)
-                      : null,
-                ),
-                child: Center(
-                  child: Text(
-                    'Почта',
-                    style: TextStyle(
-                      fontSize: 16.0,
-                      fontWeight: _selectedTab == 1
-                          ? FontWeight.w600
-                          : FontWeight.normal,
-                      color: Colors.black,
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ),
+          tab(LocaleKeys.auth_login_phone.tr(), 0),
+          const SizedBox(width: 6),
+          tab(LocaleKeys.auth_login_email.tr(), 1),
         ],
       ),
     );
   }
 
   Widget _buildPhoneInput(ThemeData theme) {
-    final dialCode = '+${_selectedCountry.phoneCode}';
+    final dialCode = '+${store.selectedCountry.phoneCode}';
     return TextField(
-      controller: _phoneController,
+      controller: phoneController,
       keyboardType: TextInputType.phone,
       inputFormatters: [
         FilteringTextInputFormatter.digitsOnly,
         LengthLimitingTextInputFormatter(15),
       ],
-      style: TextStyle(
-        fontSize: 16.0,
-        color: Colors.black,
-      ),
+      style: TextStyle(fontSize: 16.0, color: Colors.black),
       decoration: InputDecoration(
         hintText: '',
         prefixIcon: InkWell(
@@ -229,54 +189,42 @@ class _LoginScreenState extends State<LoginScreen> {
               mainAxisSize: MainAxisSize.min,
               children: [
                 Text(
-                  _selectedCountry.flagEmoji,
+                  store.selectedCountry.flagEmoji,
                   style: const TextStyle(fontSize: 20.0),
                 ),
                 const SizedBox(width: 8.0),
                 Text(
                   dialCode,
-                  style: const TextStyle(
-                    fontSize: 16.0,
-                    color: Colors.black,
-                  ),
+                  style: const TextStyle(fontSize: 16.0, color: Colors.black),
                 ),
                 const Icon(Icons.keyboard_arrow_down_rounded, size: 18.0),
               ],
             ),
           ),
         ),
-        prefixIconConstraints: const BoxConstraints(minWidth: 110.0),
-        suffixIcon: _phoneController.text.isNotEmpty
-            ? (_isPhoneValid && _phoneError == null
-                ? Padding(
-                    padding: const EdgeInsets.only(right: 16.0),
-                    child: Transform.scale(
-                      scale: 0.65,
-                      child: SvgPicture.asset(
-                        'assets/auth/accept.svg',
-                        width: 24.0,
-                        height: 24.0,
-                      ),
-                    ),
-                  )
-                : _phoneError != null
-                    ? Padding(
-                        padding: const EdgeInsets.only(right: 16.0),
-                        child: Transform.scale(
-                          scale: 0.65,
-                          child: SvgPicture.asset(
-                            'assets/auth/error.svg',
-                            width: 24.0,
-                            height: 24.0,
-                          ),
-                        ),
-                      )
-                    : null)
+        suffixIcon: store.isPhoneValid && !store.hasError
+            ? Padding(
+                padding: const EdgeInsets.only(right: 16.0),
+                child: Transform.scale(
+                  scale: 0.65,
+                  child: Assets.auth.acceptSvg.svg(),
+                ),
+              )
+            : store.hasError
+            ? Padding(
+                padding: const EdgeInsets.only(right: 16.0),
+                child: Transform.scale(
+                  scale: 0.65,
+                  child: Assets.auth.errorSvg.svg(),
+                ),
+              )
             : null,
         filled: true,
         fillColor: const Color(0xFFF0F1F3),
-        contentPadding:
-            const EdgeInsets.symmetric(horizontal: 16.0, vertical: 16.0),
+        contentPadding: const EdgeInsets.symmetric(
+          horizontal: 16.0,
+          vertical: 16.0,
+        ),
         border: OutlineInputBorder(
           borderRadius: BorderRadius.circular(8.0),
           borderSide: BorderSide.none,
@@ -284,15 +232,14 @@ class _LoginScreenState extends State<LoginScreen> {
         enabledBorder: OutlineInputBorder(
           borderRadius: BorderRadius.circular(8.0),
           borderSide: BorderSide(
-            color: _phoneError != null ? Colors.red : Colors.transparent,
+            color: store.hasError ? Colors.red : Colors.transparent,
             width: 1.0,
           ),
         ),
         focusedBorder: OutlineInputBorder(
           borderRadius: BorderRadius.circular(8.0),
           borderSide: BorderSide(
-            color:
-                _phoneError != null ? Colors.red : theme.colorScheme.secondary,
+            color: store.hasError ? Colors.red : theme.colorScheme.secondary,
             width: 2.0,
           ),
         ),
@@ -306,11 +253,8 @@ class _LoginScreenState extends State<LoginScreen> {
         ),
       ),
       onChanged: (value) {
-        setState(() {
-          final digits = value.trim();
-          _isPhoneValid = digits.length >= 5 && digits.length <= 15;
-          _phoneError = null;
-        });
+        final digits = value.trim();
+        store.phone = digits;
       },
     );
   }
@@ -322,37 +266,29 @@ class _LoginScreenState extends State<LoginScreen> {
       favorite: const ['RU', 'UA', 'KZ', 'BY'],
       countryListTheme: CountryListThemeData(
         borderRadius: BorderRadius.circular(12),
-        inputDecoration: const InputDecoration(
-          hintText: 'Поиск страны',
+        inputDecoration: InputDecoration(
+          hintText: LocaleKeys.auth_login_search_country.tr(),
           prefixIcon: Icon(Icons.search),
         ),
       ),
-      onSelect: (country) {
-        setState(() {
-          _selectedCountry = country;
-        });
-      },
+      onSelect: (country) => store.setSelectedCountry(country),
     );
   }
 
   Widget _buildEmailInput(ThemeData theme) {
     return TextField(
-      controller: _emailController,
+      controller: emailController,
       keyboardType: TextInputType.emailAddress,
-      style: TextStyle(
-        fontSize: 16.0,
-        color: Colors.black,
-      ),
+      style: TextStyle(fontSize: 16.0, color: Colors.black),
       decoration: InputDecoration(
-        hintText: 'Введите почту',
-        hintStyle: TextStyle(
-          color: Colors.grey[400],
-          fontSize: 16.0,
-        ),
+        hintText: LocaleKeys.auth_login_enter_email.tr(),
+        hintStyle: TextStyle(color: Colors.grey[400], fontSize: 16.0),
         filled: true,
         fillColor: const Color(0xFFF0F1F3),
-        contentPadding:
-            const EdgeInsets.symmetric(horizontal: 16.0, vertical: 16.0),
+        contentPadding: const EdgeInsets.symmetric(
+          horizontal: 16.0,
+          vertical: 16.0,
+        ),
         border: OutlineInputBorder(
           borderRadius: BorderRadius.circular(8.0),
           borderSide: BorderSide.none,
@@ -360,15 +296,14 @@ class _LoginScreenState extends State<LoginScreen> {
         enabledBorder: OutlineInputBorder(
           borderRadius: BorderRadius.circular(8.0),
           borderSide: BorderSide(
-            color: _emailError != null ? Colors.red : Colors.transparent,
+            color: store.hasError ? Colors.red : Colors.transparent,
             width: 1.0,
           ),
         ),
         focusedBorder: OutlineInputBorder(
           borderRadius: BorderRadius.circular(8.0),
           borderSide: BorderSide(
-            color:
-                _emailError != null ? Colors.red : theme.colorScheme.secondary,
+            color: store.hasError ? Colors.red : theme.colorScheme.secondary,
             width: 2.0,
           ),
         ),
@@ -381,76 +316,49 @@ class _LoginScreenState extends State<LoginScreen> {
           borderSide: BorderSide(color: Colors.red, width: 2.0),
         ),
       ),
-      onChanged: (value) {
-        setState(() {
-          _isEmailValid = value.contains('@') && value.contains('.');
-          if (value.isNotEmpty && !_isEmailValid) {
-            _emailError = 'Неверный формат email';
-          } else {
-            _emailError = null;
-          }
-        });
-      },
+      onChanged: (value) => store.email = value,
     );
   }
 
   Widget _buildNextButton(ThemeData theme) {
-    final isValid = _selectedTab == 0 ? _isPhoneValid : _isEmailValid;
-    final hasError = (_selectedTab == 0 && _phoneError != null) ||
-        (_selectedTab == 1 && _emailError != null);
-
     return SizedBox(
       width: double.infinity,
       child: ElevatedButton(
-        onPressed: (isValid && !hasError && !_isSendingCode)
-            ? () async {
-                if (_selectedTab == 0) {
-                  final digits = _phoneController.text.trim();
-                  final phoneNumber = '+${_selectedCountry.phoneCode}$digits';
-                  setState(() {
-                    _isSendingCode = true;
-                    _phoneError = null;
-                  });
-                  try {
-                    final authStore = locator<AuthStore>();
-                    await authStore.sendVerificationCode(phoneNumber);
-                    if (!mounted) return;
-                    context.push('/auth/verification?phoneNumber=$phoneNumber');
-                  } catch (e) {
-                    if (!mounted) return;
-                    setState(() {
-                      _phoneError =
-                          'Не удалось отправить код. Попробуйте позже.';
-                    });
-                    ScaffoldMessenger.of(context).showSnackBar(
-                      const SnackBar(
-                        content:
-                            Text('Ошибка отправки кода. Повторите попытку.'),
-                      ),
-                    );
-                  } finally {
-                    if (mounted) {
-                      setState(() {
-                        _isSendingCode = false;
-                      });
+        onPressed: (store.canSend)
+            ? () {
+                if (store.selectedTab == 0) {
+                  store.sendVerificationCode().then((result) {
+                    if (result && mounted) {
+                      context.push(
+                        '/auth/verification?phoneNumber=${store.phoneNumber}',
+                      );
+                    } else {
+                      if (mounted) {
+                        ScaffoldMessenger.of(context).showSnackBar(
+                          SnackBar(
+                            content: Text(
+                              LocaleKeys.auth_login_error_send_code.tr(),
+                            ),
+                          ),
+                        );
+                      }
                     }
-                  }
+                  });
                 }
               }
             : null,
         style: ElevatedButton.styleFrom(
-          backgroundColor: isValid && !hasError
+          backgroundColor: store.canSend
               ? AppColors.primary
               : const Color(0xFFF0F1F3),
-          foregroundColor:
-              isValid && !hasError ? Colors.white : Colors.grey[600],
+          foregroundColor: store.canSend ? Colors.white : Colors.grey[600],
           padding: const EdgeInsets.symmetric(vertical: 16.0),
           shape: RoundedRectangleBorder(
             borderRadius: BorderRadius.circular(8.0),
           ),
           elevation: 0,
         ),
-        child: _isSendingCode
+        child: store.isLoading
             ? const SizedBox(
                 height: 20.0,
                 width: 20.0,
@@ -460,11 +368,8 @@ class _LoginScreenState extends State<LoginScreen> {
                 ),
               )
             : Text(
-                'Далее',
-                style: TextStyle(
-                  fontSize: 16.0,
-                  fontWeight: FontWeight.w500,
-                ),
+                LocaleKeys.auth_login_button_title.tr(),
+                style: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w500),
               ),
       ),
     );
@@ -474,61 +379,47 @@ class _LoginScreenState extends State<LoginScreen> {
     return Column(
       children: [
         Text(
-          'Или войти с помощью',
+          LocaleKeys.auth_login_or_login_with.tr(),
           style: TextStyle(
-            fontSize: 14.0,
-            color: Colors.grey[600],
+            fontFamily: FontFamily.inter,
+            fontSize: 13.0,
+            color: AppColors.onSubSurface,
+            fontWeight: FontWeight.w500,
           ),
         ),
-        const SizedBox(height: 16.0),
+        const SizedBox(height: 4.0),
         Row(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            _buildSocialButton('assets/auth/google.png', false, () {}),
-            const SizedBox(width: 16.0),
-            _buildSocialButton('assets/auth/tg_white.png', true, () {}),
-            const SizedBox(width: 16.0),
-            _buildSocialButton('assets/auth/apple.png', false, () {}),
+            _buildSocialButton(Assets.icons.google.svg(), false, () {}),
+            const SizedBox(width: 12.0),
+            _buildSocialButton(Assets.icons.telegram.svg(), true, () {}),
+            const SizedBox(width: 12.0),
+            _buildSocialButton(Assets.icons.apple.svg(), false, () {}),
           ],
         ),
       ],
     );
   }
 
-  Widget _buildSocialButton(
-      String assetPath, bool isTelegram, VoidCallback onTap) {
+  Widget _buildSocialButton(Widget icon, bool isTelegram, VoidCallback onTap) {
     return GestureDetector(
       onTap: onTap,
       child: Container(
-        width: 40.0,
-        height: 40.0,
+        width: 32.0,
+        height: 32.0,
         decoration: BoxDecoration(
           gradient: isTelegram
               ? LinearGradient(
-                  colors: [
-                    Color(0xFF2AABEE),
-                    Color(0xFF229ED9),
-                  ],
+                  colors: [Color(0xFF2AABEE), Color(0xFF229ED9)],
                   begin: Alignment.topLeft,
                   end: Alignment.bottomRight,
                 )
               : null,
           color: isTelegram ? null : Colors.white,
           borderRadius: BorderRadius.circular(8.0),
-          border: isTelegram
-              ? null
-              : Border.all(color: Colors.grey[300]!, width: 1.0),
         ),
-        child: Center(
-          child: Image.asset(
-            assetPath,
-            width: 32.0,
-            height: 32.0,
-            errorBuilder: (context, error, stackTrace) {
-              return Icon(Icons.image, size: 32.0);
-            },
-          ),
-        ),
+        child: Center(child: icon),
       ),
     );
   }
