@@ -1,150 +1,11 @@
 import 'package:dio/dio.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:injectable/injectable.dart';
 
-class ApiProvider {
+@singleton
+final class ApiProvider {
   late final Dio _dio;
-  final String baseUrl;
-  static const _skipAuthKey = 'skipAuth';
 
-  ApiProvider({
-    required this.baseUrl,
-  }) {
-    _dio = Dio(
-      BaseOptions(
-        baseUrl: baseUrl,
-        connectTimeout: const Duration(seconds: 30),
-        receiveTimeout: const Duration(seconds: 30),
-        headers: {
-          'Content-Type': 'application/json',
-          'Accept': 'application/json',
-        },
-      ),
-    );
-
-    _setupInterceptors();
-  }
-
-  void _setupInterceptors() {
-    _dio.interceptors.add(
-      InterceptorsWrapper(
-        onRequest: (options, handler) async {
-          final skipAuth = options.extra[_skipAuthKey] == true;
-          if (!skipAuth) {
-            final token = await _getAccessToken();
-            if (token != null) {
-              options.headers['Authorization'] = 'Bearer $token';
-            }
-          }
-          return handler.next(options);
-        },
-        onResponse: (response, handler) {
-          return handler.next(response);
-        },
-        onError: (error, handler) async {
-          final skipAuth = error.requestOptions.extra[_skipAuthKey] == true;
-          if (skipAuth) {
-            return handler.next(error);
-          }
-          if (error.response?.statusCode == 401) {
-            final refreshed = await _refreshToken();
-            if (refreshed) {
-              final opts = error.requestOptions;
-              final token = await _getAccessToken();
-              if (token != null) {
-                opts.headers['Authorization'] = 'Bearer $token';
-                try {
-                  final response = await _dio.fetch(opts);
-                  return handler.resolve(response);
-                } catch (e) {
-                  return handler.next(error);
-                }
-              }
-            }
-            await _clearTokens();
-          }
-          return handler.next(error);
-        },
-      ),
-    );
-
-    _dio.interceptors.add(
-      LogInterceptor(
-        requestBody: true,
-        responseBody: true,
-        error: true,
-      ),
-    );
-  }
-
-  Future<String?> _getAccessToken() async {
-    final prefs = await SharedPreferences.getInstance();
-    return prefs.getString('access_token');
-  }
-
-  Future<String?> _getRefreshToken() async {
-    final prefs = await SharedPreferences.getInstance();
-    return prefs.getString('refresh_token');
-  }
-
-  Future<bool> _refreshToken() async {
-    try {
-      final refreshToken = await _getRefreshToken();
-      if (refreshToken == null) return false;
-
-      final response = await _dio.post(
-        '/api/auth/refresh',
-        data: {'refresh_token': refreshToken},
-        options: Options(
-          extra: const {_skipAuthKey: true},
-        ),
-      );
-
-      if (response.statusCode == 200) {
-        final data = _extractData(response.data);
-        final newAccessToken = data['access_token'] as String? ?? '';
-        if (newAccessToken.isEmpty) {
-          return false;
-        }
-        await _saveTokens(
-          accessToken: newAccessToken,
-          refreshToken: data['refresh_token'] as String? ?? refreshToken,
-        );
-        return true;
-      }
-      return false;
-    } catch (e) {
-      return false;
-    }
-  }
-
-  Future<void> _saveTokens({
-    required String accessToken,
-    required String refreshToken,
-  }) async {
-    final prefs = await SharedPreferences.getInstance();
-    await prefs.setString('access_token', accessToken);
-    await prefs.setString('refresh_token', refreshToken);
-  }
-
-  Future<void> _clearTokens() async {
-    final prefs = await SharedPreferences.getInstance();
-    await prefs.remove('access_token');
-    await prefs.remove('refresh_token');
-  }
-
-  Future<void> setTokens({
-    required String accessToken,
-    required String refreshToken,
-  }) async {
-    await _saveTokens(
-      accessToken: accessToken,
-      refreshToken: refreshToken,
-    );
-  }
-
-  Future<void> clearTokens() async {
-    await _clearTokens();
-  }
+  ApiProvider(this._dio);
 
   Future<Response<T>> get<T>(
     String path, {
@@ -199,16 +60,4 @@ class ApiProvider {
       options: options,
     );
   }
-
-  Map<String, dynamic> _extractData(dynamic raw) {
-    if (raw is Map<String, dynamic>) {
-      final data = raw['data'];
-      if (data is Map<String, dynamic>) {
-        return data;
-      }
-      return raw;
-    }
-    return <String, dynamic>{};
-  }
 }
-
