@@ -1,6 +1,6 @@
+import 'package:exchats/core/services/secure_storage/token_repository.dart';
+import 'package:exchats/features/splash/presentation/ui/splash_screen.dart';
 import 'package:go_router/go_router.dart';
-import '../../../../core/di/locator.dart';
-import '../../features/auth/presentation/store/auth_store.dart';
 import '../../features/auth/presentation/ui/auth/auth_screen.dart';
 import '../../features/auth/presentation/ui/auth/login/login_screen.dart';
 import '../../features/auth/presentation/ui/auth/verification/verification_screen.dart';
@@ -24,22 +24,17 @@ import '../../features/call/presentation/ui/calls/new_call_screen.dart';
 
 class AppRouter {
   static final GoRouter router = GoRouter(
-    initialLocation: '/auth',
-    redirect: (context, state) {
-      final authStore = locator<AuthStore>();
-      final isAuthRoute = state.matchedLocation.startsWith('/auth');
-      
-      if (!authStore.isAuthenticated && !isAuthRoute) {
-        return '/auth';
-      }
-      
-      if (authStore.isAuthenticated && isAuthRoute) {
-        return '/';
-      }
-      
-      return null;
-    },
+    initialLocation: '/splash',
     routes: [
+      GoRoute(
+        path: '/splash',
+        name: 'splash ',
+        builder: (context, state) => const SplashScreen(),
+        redirect: (context, state) async {
+          final isAuthenticated = await TokenRepository.hasAccessToken();
+          return isAuthenticated ? '/' : '/auth';
+        },
+      ),
       GoRoute(
         path: '/auth',
         builder: (context, state) => const AuthScreen(),
@@ -53,7 +48,8 @@ class AppRouter {
             path: 'verification',
             name: 'verification',
             builder: (context, state) {
-              final phoneNumber = state.uri.queryParameters['phoneNumber'] ?? '';
+              final phoneNumber =
+                  state.uri.queryParameters['phoneNumber'] ?? '';
               return VerificationScreen(phoneNumber: phoneNumber);
             },
           ),
@@ -98,8 +94,10 @@ class AppRouter {
             builder: (context, state) {
               final userId = state.uri.queryParameters['userId'];
               final userName = state.uri.queryParameters['userName'];
-              final isIncoming = state.uri.queryParameters['isIncoming'] == 'true';
-              final isVideoCall = state.uri.queryParameters['isVideoCall'] == 'true';
+              final isIncoming =
+                  state.uri.queryParameters['isIncoming'] == 'true';
+              final isVideoCall =
+                  state.uri.queryParameters['isVideoCall'] == 'true';
               return CallScreen(
                 userId: userId,
                 userName: userName,
@@ -144,7 +142,8 @@ class AppRouter {
             builder: (context, state) {
               final userId = state.uri.queryParameters['userId'] ?? '';
               final userName = state.uri.queryParameters['userName'] ?? '';
-              final userStatus = state.uri.queryParameters['userStatus'] ?? 'offline';
+              final userStatus =
+                  state.uri.queryParameters['userStatus'] ?? 'offline';
               return UserProfileScreen(
                 userId: userId,
                 userName: userName,
@@ -157,11 +156,9 @@ class AppRouter {
             name: 'group_profile',
             builder: (context, state) {
               final groupId = state.uri.queryParameters['groupId'] ?? '';
-              final groupName = state.uri.queryParameters['groupName'] ?? 'Группа';
-              return GroupProfileScreen(
-                groupId: groupId,
-                groupName: groupName,
-              );
+              final groupName =
+                  state.uri.queryParameters['groupName'] ?? 'Группа';
+              return GroupProfileScreen(groupId: groupId, groupName: groupName);
             },
           ),
           GoRoute(
@@ -170,7 +167,8 @@ class AppRouter {
             builder: (context, state) {
               final userId = state.uri.queryParameters['userId'] ?? '';
               final userName = state.uri.queryParameters['userName'] ?? '';
-              final isVideoCall = state.uri.queryParameters['isVideoCall'] == 'true';
+              final isVideoCall =
+                  state.uri.queryParameters['isVideoCall'] == 'true';
               return ActiveCallScreen(
                 userId: userId,
                 userName: userName,
