@@ -1,9 +1,10 @@
-class RefreshRequestDto {
-  final String refreshToken;
+import 'package:freezed_annotation/freezed_annotation.dart';
 
-  const RefreshRequestDto({required this.refreshToken});
+part 'refresh_request_dto.freezed.dart';
+part 'refresh_request_dto.g.dart';
 
-  Map<String, dynamic> toJson() {
-    return {'refresh_token': refreshToken};
-  }
+@Freezed(copyWith: false, equal: false, fromJson: false, toJson: true)
+abstract class RefreshRequestDto with _$RefreshRequestDto {
+  const factory RefreshRequestDto({required String refreshToken}) =
+      _RefreshRequestDto;
 }
