@@ -1,9 +1,10 @@
-class RegisterRequestDto {
-  final String phone;
+import 'package:freezed_annotation/freezed_annotation.dart';
 
-  const RegisterRequestDto({required this.phone});
+part 'register_request_dto.freezed.dart';
+part 'register_request_dto.g.dart';
 
-  Map<String, dynamic> toJson() {
-    return {'phone': phone};
-  }
+@Freezed(copyWith: false, equal: false, fromJson: false, toJson: true)
+abstract class RegisterRequestDto with _$RegisterRequestDto {
+  const factory RegisterRequestDto({required String phone}) =
+      _RegisterRequestDto;
 }
