@@ -1,16 +1,12 @@
-class VerifyRequestDto {
-  final String phone;
-  final String code;
+import 'package:freezed_annotation/freezed_annotation.dart';
 
-  const VerifyRequestDto({
-    required this.phone,
-    required this.code,
-  });
+part 'verify_request_dto.freezed.dart';
+part 'verify_request_dto.g.dart';
 
-  Map<String, dynamic> toJson() {
-    return {
-      'phone': phone,
-      'code': code,
-    };
-  }
+@Freezed(copyWith: false, equal: false, fromJson: false, toJson: true)
+abstract class VerifyRequestDto with _$VerifyRequestDto {
+  const factory VerifyRequestDto({
+    required String phone,
+    required String code,
+  }) = _VerifyRequestDto;
 }
