@@ -10,7 +10,7 @@ class ArchiveScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: AppColors.background,
+      backgroundColor: AppColors.surface,
       appBar: AppBar(
         backgroundColor: AppColors.white,
         elevation: 0,
@@ -31,10 +31,7 @@ class ArchiveScreen extends StatelessWidget {
       body: Center(
         child: Text(
           AppStrings.archiveEmpty,
-          style: TextStyle(
-            color: AppColors.grey(600),
-            fontSize: 16.0,
-          ),
+          style: TextStyle(color: AppColors.grey(600), fontSize: 16.0),
         ),
       ),
     );
